+++ src/app/(afterLogin)/(profile)/family-edit/family-delete/page.tsx
@@ -1,25 +1,7 @@
 'use client';
 
-import Modal from '@/app/_components/common/modal/Modal';
-import { FAMILY_DELETE_CONTENT } from '@/app/_constants/modal';
-import { useRouter } from 'next/navigation';
+import FamilyDelete from '@/app/(afterLogin)/(profile)/family-edit/_components/FamilyDelete';
 
-export default function FamilyDelete() {
-  const router = useRouter();
-  const handleCancelClick = () => {
-    router.back();
-  };
-
-  const handleDeletionConfirm = () => {};
-  return (
-    <div className="absolute">
-      <Modal
-        title="가족 삭제하기"
-        content={FAMILY_DELETE_CONTENT}
-        buttonText="가족 삭제"
-        onCancel={handleCancelClick}
-        onExecute={handleDeletionConfirm}
-      />
-    </div>
-  );
+export default function FamilyDeletePage() {
+  return <FamilyDelete />;
 }
