+++ src/app/_api/apiRoutes.ts
@@ -1,7 +1,11 @@
 export const apiRoutes = {
+  // Family
   createFamily: '/v1/groups/create',
   UpdateFamily: '/v1/groups/update',
   getFamilyList: '/v1/groups/my',
+  deleteFamily: '/v1/groups/delete',
+
+  // User
   UpdateProfile: '/v1/users/profile',
   UserLogout: '/v1/users/logout',
   UserDeletion: '/v1/users/delete',
